Made fonts consistent across the app. All are now in sans-serif to match synchrony brand
