+++ MyApplication/app/src/main/AndroidManifest.xml
@@ -31,7 +31,7 @@
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
         <activity
-            android:name=".directory.Profile"
+            android:name=".housing.HousingFragment$Profile"
             android:label="@string/title_activity_main"
             android:theme="@style/AppTheme.NoActionBar" />
     </application>
