+++ MyApplication/app/src/main/res/layout/activity_housing.xml
@@ -14,12 +14,14 @@
         android:paddingTop="10dp"
         android:paddingBottom="10dp"
         android:id="@+id/apartmentName"
+        android:fontFamily="sans-serif"
         />
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/my_recycler_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:scrollbars="vertical" />
+        android:scrollbars="vertical"
+        android:fontFamily="sans-serif"/>
 
 </LinearLayout>
