+++ MyApplication/app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,8 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:openDrawer="start">
+    tools:openDrawer="start"
+    android:fontFamily="sans-serif">
 
     <include
         layout="@layout/app_bar_main"
@@ -20,6 +21,7 @@
         android:layout_gravity="start"
         android:fitsSystemWindows="true"
         app:headerLayout="@layout/nav_header_main"
-        app:menu="@menu/activity_main_drawer" />
+        app:menu="@menu/activity_main_drawer"
+        android:fontFamily="sans-serif"/>
 
 </android.support.v4.widget.DrawerLayout>
