+++ MyApplication/app/src/main/res/layout/app_bar_main.xml
@@ -5,19 +5,22 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.syfblp.sas.blpappv2.MainActivity">
+    tools:context="com.syfblp.sas.blpappv2.MainActivity"
+    android:fontFamily="sans-serif">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:theme="@style/AppTheme.AppBarOverlay">
+        android:theme="@style/AppTheme.AppBarOverlay"
+        android:fontFamily="sans-serif">
 
         <android.support.v7.widget.Toolbar
             android:id="@+id/toolbar"
             android:layout_width="match_parent"
             android:layout_height="?attr/actionBarSize"
             android:background="?attr/colorPrimary"
-            app:popupTheme="@style/AppTheme.PopupOverlay" />
+            app:popupTheme="@style/AppTheme.PopupOverlay"
+            android:fontFamily="sans-serif"/>
 
     </android.support.design.widget.AppBarLayout>
