+++ MyApplication/app/src/main/res/layout/content_blpdirectory.xml
@@ -8,12 +8,14 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
+    android:fontFamily="sans-serif"
 >
 
 
     <ListView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:id="@+id/BLPDirectorylistView" />
+        android:id="@+id/BLPDirectorylistView"
+        android:fontFamily="sans-serif"/>
 
 </RelativeLayout>
