+++ MyApplication/app/src/main/res/layout/content_main.xml
@@ -10,7 +10,8 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:context="com.syfblp.sas.blpappv2.MainActivity"
-    tools:showIn="@layout/app_bar_main">
+    tools:showIn="@layout/app_bar_main"
+    android:fontFamily="sans-serif">
 
     <FrameLayout
     android:layout_width="match_parent"
