+++ MyApplication/app/src/main/res/layout/content_profile.xml
@@ -8,6 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
+    android:fontFamily="sans-serif"
     >
 
     <TextView
@@ -18,7 +19,8 @@
         android:layout_alignParentTop="true"
         android:layout_centerHorizontal="true"
         android:editable="false"
-        android:textSize="@dimen/abc_text_size_headline_material" />
+        android:textSize="@dimen/abc_text_size_headline_material"
+        android:fontFamily="sans-serif"/>
 
     <TextView
         android:layout_width="wrap_content"
@@ -27,7 +29,8 @@
         android:text="@string/FunctionLocation"
         android:id="@+id/txtFunctionLocation"
         android:layout_below="@+id/txtFirstLastName"
-        android:layout_centerHorizontal="true"/>
+        android:layout_centerHorizontal="true"
+        android:fontFamily="sans-serif"/>
 
     <TextView
         android:layout_width="wrap_content"
@@ -37,7 +40,10 @@
         android:id="@+id/txtRoleLbl"
         android:layout_below="@+id/txtFunctionLocation"
         android:layout_alignParentStart="true"
-        android:layout_marginTop="50dp"/>
+        android:layout_marginTop="50dp"
+        android:fontFamily="sans-serif"
+        android:textStyle="bold"
+        android:textSize="17dp" />
 
     <TextView
         android:layout_width="wrap_content"
@@ -46,7 +52,9 @@
         android:text="@string/role"
         android:id="@+id/txtRole"
         android:layout_below="@+id/txtRoleLbl"
-        android:layout_alignParentStart="true"/>
+        android:layout_alignParentStart="true"
+        android:fontFamily="sans-serif"
+        android:textSize="14dp" />
 
     <TextView
         android:layout_width="wrap_content"
@@ -56,7 +64,9 @@
         android:id="@+id/txtAL"
         android:layout_alignParentStart="true"
         android:layout_below="@+id/txtALLbl"
-        android:layout_alignParentLeft="true"/>
+        android:layout_alignParentLeft="true"
+        android:fontFamily="sans-serif"
+        android:textSize="14dp" />
 
     <TextView
         android:layout_width="wrap_content"
@@ -66,7 +76,10 @@
         android:id="@+id/txtALLbl"
         android:layout_below="@+id/txtRoleLbl"
         android:layout_alignParentStart="true"
-        android:layout_marginTop="43dp" />
+        android:layout_marginTop="43dp"
+        android:fontFamily="sans-serif"
+        android:textStyle="bold"
+        android:textSize="17dp" />
 
     <TextView
         android:layout_width="wrap_content"
@@ -76,7 +89,10 @@
         android:id="@+id/txtPhone"
         android:layout_below="@+id/txtALLbl"
         android:layout_alignParentStart="true"
-        android:layout_marginTop="58dp" />
+        android:layout_marginTop="58dp"
+        android:fontFamily="sans-serif"
+        android:textStyle="bold"
+        android:textSize="17dp" />
 
     <TextView
         android:layout_width="wrap_content"
@@ -85,6 +101,9 @@
         android:text="@string/email"
         android:id="@+id/txtEmail"
         android:layout_below="@+id/txtPhone"
-        android:layout_alignParentStart="true" />
+        android:layout_alignParentStart="true"
+        android:fontFamily="sans-serif"
+        android:textStyle="bold"
+        android:textSize="17dp" />
 
 </RelativeLayout>
