+++ MyApplication/app/src/main/res/layout/event_fragment.xml
@@ -2,7 +2,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="${relativePackage}.${activityClass}">
+    tools:context="${relativePackage}.${activityClass}"
+    android:fontFamily="sans-serif">
 
     <!-- A RecyclerView with some commonly used attributes -->
     <android.support.v7.widget.RecyclerView
@@ -10,6 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:scrollbars="vertical"
-        android:layout_alignParentStart="true"/>
+        android:layout_alignParentStart="true"
+        android:fontFamily="sans-serif"/>
 
 </RelativeLayout>
