+++ MyApplication/app/src/main/res/layout/fragment_event.xml
@@ -3,6 +3,7 @@
     android:layout_width="fill_parent"
     android:layout_height="fill_parent"
     android:orientation="vertical"
-    android:background="#f4f4f4" >
+    android:background="#f4f4f4"
+    android:fontFamily="sans-serif">
 
 </LinearLayout>
