+++ MyApplication/app/src/main/res/layout/fragment_housing.xml
@@ -1,11 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:fontFamily="sans-serif">
 
     <ListView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:id="@+id/listView"
-        android:choiceMode="none" />
+        android:choiceMode="none"
+        android:fontFamily="sans-serif"/>
 </LinearLayout>
