+++ MyApplication/app/src/main/res/layout/fragment_import.xml
@@ -17,6 +17,7 @@
         android:layout_height="wrap_content"
         android:id="@+id/BLPDirectorylistView"
         android:layout_alignParentTop="true"
-        android:layout_alignParentStart="true" />
+        android:layout_alignParentStart="true"
+        android:fontFamily="sans-serif"/>
 
 </RelativeLayout>
