+++ MyApplication/app/src/main/res/layout/housing_details.xml
@@ -3,7 +3,8 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:fontFamily="sans-serif">
 
     <TextView
         android:layout_width="wrap_content"
@@ -13,12 +14,14 @@
         android:paddingLeft="10dp"
         android:paddingTop="10dp"
         android:paddingBottom="10dp"
-        android:id="@+id/apartmentName"/>
+        android:id="@+id/apartmentName"
+        android:fontFamily="sans-serif"/>
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/my_recycler_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:scrollbars="vertical" />
+        android:scrollbars="vertical"
+        android:fontFamily="sans-serif"/>
 
 </LinearLayout>
