+++ MyApplication/app/src/main/res/layout/list_group.xml
@@ -4,7 +4,8 @@
     android:layout_height="wrap_content"
     android:orientation="vertical"
     android:padding="8dp" 
-    android:background="#000000">
+    android:background="#000000"
+    android:fontFamily="sans-serif">
 
 
     <TextView
@@ -13,6 +14,7 @@
         android:layout_height="wrap_content"
         android:paddingLeft="?android:attr/expandableListPreferredItemPaddingLeft"
         android:textSize="17dp"
-        android:textColor="#f9f93d" />
+        android:textColor="#f9f93d"
+        android:fontFamily="sans-serif"/>
 
 </LinearLayout>
